Swap the two variables
Can you think of a way to swap the values
of two variables that does not need a third variable
as a temporary storage? In the code block below, try
to implement the swapping of the values of a and b
without using a third variable. To help you out, the
first step to do this is already given. You just need
to add two more lines of code.